Removed the payment methoods abstract class
