+++ Model/Methods/ApplePayMethod.php
@@ -19,7 +19,7 @@ namespace CheckoutCom\Magento2\Model\Methods;
 
 use CheckoutCom\Magento2\Gateway\Config\Config;
 
-class ApplePayMethod extends Method
+class ApplePayMethod extends \Magento\Payment\Model\Method\AbstractMethod
 {
     /**
      * @var string
@@ -32,6 +32,67 @@ class ApplePayMethod extends Method
      */
     protected $_code = self::CODE;
 
+    public function __construct(
+        \Magento\Framework\Model\Context $context,
+        \Magento\Framework\Registry $registry,
+        \Magento\Framework\Api\ExtensionAttributesFactory $extensionFactory,
+        \Magento\Framework\Api\AttributeValueFactory $customAttributeFactory,
+        \Magento\Payment\Helper\Data $paymentData,
+        \Magento\Framework\App\Config\ScopeConfigInterface $scopeConfig,
+        \Magento\Payment\Model\Method\Logger $logger,
+        \Magento\Backend\Model\Auth\Session $backendAuthSession,
+        \Magento\Checkout\Model\Cart $cart,
+        \Magento\Framework\UrlInterface $urlBuilder,
+        \Magento\Framework\ObjectManagerInterface $objectManager,
+        \Magento\Sales\Model\Order\Email\Sender\InvoiceSender $invoiceSender,
+        \Magento\Framework\DB\TransactionFactory $transactionFactory,
+        \Magento\Customer\Model\Session $customerSession,
+        \Magento\Checkout\Model\Session $checkoutSession,
+        \Magento\Checkout\Helper\Data $checkoutData,
+        \Magento\Quote\Api\CartRepositoryInterface $quoteRepository,
+        \Magento\Quote\Api\CartManagementInterface $quoteManagement,
+        \Magento\Sales\Model\Order\Email\Sender\OrderSender $orderSender,
+        \Magento\Backend\Model\Session\Quote $sessionQuote,
+        \Magento\Framework\HTTP\PhpEnvironment\RemoteAddress $remoteAddress,
+        \CheckoutCom\Magento2\Gateway\Config\Config $config,
+        \CheckoutCom\Magento2\Model\Service\apiHandlerService $apiHandler,
+        \CheckoutCom\Magento2\Model\Service\QuoteHandlerService $quoteHandler,
+        \Magento\Framework\Model\ResourceModel\AbstractResource $resource = null,
+        \Magento\Framework\Data\Collection\AbstractDb $resourceCollection = null,
+        array $data = []
+    ) {
+        parent::__construct(
+            $context,
+            $registry,
+            $extensionFactory,
+            $customAttributeFactory,
+            $paymentData,
+            $scopeConfig,
+            $logger,
+            $resource,
+            $resourceCollection,
+            $data
+        );
+
+        $this->urlBuilder         = $urlBuilder;
+        $this->backendAuthSession = $backendAuthSession;
+        $this->cart               = $cart;
+        $this->_objectManager     = $objectManager;
+        $this->invoiceSender      = $invoiceSender;
+        $this->transactionFactory = $transactionFactory;
+        $this->customerSession    = $customerSession;
+        $this->checkoutSession    = $checkoutSession;
+        $this->checkoutData       = $checkoutData;
+        $this->quoteRepository    = $quoteRepository;
+        $this->quoteManagement    = $quoteManagement;
+        $this->orderSender        = $orderSender;
+        $this->sessionQuote       = $sessionQuote;
+        $this->remoteAddress      = $remoteAddress;
+        $this->config             = $config;
+        $this->apiHandler         = $apiHandler;
+        $this->quoteHandler       = $quoteHandler;
+    }
+    
     /**
      * { function_description }
      *
